inpainting-in-medical-imaging-egemen-dev/requirements.txt
torchvision
tensorboard
segmentation-models-pytorch
kornia
opencv-python
scikit-learn
pandas
numpy
nibabel
albumentations
notebook
tqdm
matplotlib
seaborn
focal-frequency-loss
sewar
piq